services/requirements.txt
fastapi
uvicorn[standard]
detoxify
transformers
torch
torchvision
Pillow
httpx
python-multipart
pydantic
pytest
